Improve xdebug-on/xdebug-off when already on/off

+++ bin/xdebug-off.sh
@@ -5,5 +5,11 @@ set -e
 if [ -e /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini ]; then
   rm /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini
   pkill -o -USR2 php-fpm
+  echo "Xdebug disabled"
+else
+  echo "Xdebug already disabled"
 fi
-rm /app/.xdebug-enabled
+
+if [ -e /app/.xdebug-enabled ]; then
+  rm /app/.xdebug-enabled
+fi
